+++ Craftsman/Builders/InfrastructureServiceRegistrationBuilder.cs
@@ -26,9 +26,11 @@
         public static string GetServiceRegistrationText(string srcDirectory, string projectBaseName, string classNamespace)
         {
             var dbContextClassPath = ClassPathHelper.DbContextClassPath(srcDirectory, "", projectBaseName);
+            var utilsClassPath = ClassPathHelper.WebApiUtilsClassPath(srcDirectory, "", projectBaseName);
             return @$"namespace {classNamespace}
 {{
     using {dbContextClassPath.ClassNamespace};
+    using {utilsClassPath.ClassNamespace};
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.EntityFrameworkCore;
     using Microsoft.Extensions.Configuration;